Refactored ajax search:  created SearchController and made comments to search.js.
