+++ src/main/java/com/givehopeweb/repositories/Charities.java
@@ -12,9 +12,7 @@ import java.util.List;
 public interface Charities extends CrudRepository <Charity, Integer> {
 
         @Query ("select c from Charity c where c.user.id = ?1")
-
         List<Charity> findUserFavorites (int userId);
-
 
         List<Charity>
         findByCharityNameContainingOrCategoryContainingOrStateContainingOrCityContainingOrDescriptionContaining
